fix & feat
fix: rmdir can delete root directory

feat: dataserver and nameserver registration

refactor: dataserver api mod, change some PUT method to POST, client also changed

refactor: remove pkg/client/api

docs: update client/dataserver/nameserver design

build: add make demo.reset to reset demo data

build: set `all` as first target

todo: to finish client demo and nameserver client

todo: to finish health keeping

+++ pkg/nameserver/apiserver/blob/v1/service/rmdir.go
@@ -6,13 +6,20 @@ import (
 	"go-dfs-server/pkg/nameserver/server"
 	"go-dfs-server/pkg/utils"
 	"os"
-	"path/filepath"
 	"sync"
 )
 
 func (b blobService) Rmdir(path string) error {
-	directoryPath := filepath.Join(server.GlobalServerDesc.Opt.Volume, path)
-	_, err := os.Stat(directoryPath)
+	directoryPath, err := utils.JoinSubPathSafe(server.GlobalServerDesc.Opt.Volume, path)
+	if err != nil {
+		return err
+	}
+
+	if utils.IsSameDirectory(server.GlobalServerDesc.Opt.Volume, directoryPath) {
+		return errors.New("cannot delete this directory")
+	}
+
+	_, err = os.Stat(directoryPath)
 	if err == nil {
 		err = os.RemoveAll(directoryPath)
 		if err != nil {
